+++ src/medipro/object/camera/SmoothFollowingCameraModel.java
@@ -5,7 +5,7 @@ import medipro.world.World;
 
 public class SmoothFollowingCameraModel extends FollowingCameraModel {
 
-    public float followingSpeed = 0.1f;
+    public Double followingSpeed = 0.1;
 
     public SmoothFollowingCameraModel(World world, GameObjectModel target) {
         super(world, target);
